pyproject.toml
[tool.poetry]
name = "olas-operate-middleware"
version = "0.2.0-rc9"
description = ""
authors = ["David Vilela <dvilelaf@gmail.com>", "Viraj Patel <vptl185@gmail.com>"]
readme = "README.md"
packages = [
    { include = "operate" }
]
include = [
    "operate/data/acn/*",
    "operate/data/contracts/*",
    "operate/data/contracts/staking_token/*",
    "operate/data/contracts/staking_token/build/*",
]

[tool.poetry.scripts]
operate = "operate.cli:main"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
open-autonomy = "==0.18.3"
open-aea-ledger-cosmos = "==1.60.0"
open-aea-ledger-ethereum = "==1.60.0"
open-aea-ledger-ethereum-flashbots = "==1.60.0"
open-aea-cli-ipfs = "==1.60.0"
clea = "==0.1.0rc4"
cytoolz = "==0.12.3"
docker = "6.1.2"
fastapi = "==0.115.6"
eth-hash = "==0.7.0"
eth-account = "==0.8.0"
eth-keyfile = "==0.6.1"
eth-keys = "==0.4.0"
eth-rlp = "==0.3.0"
eth-typing = "==3.5.2"
eth-utils = "==2.3.1"
eth-abi = "==5.1.0"
frozenlist = "==1.4.1"
hexbytes = "==0.3.1"
ipfshttpclient = "==0.8.0a2"
jsonschema = "==4.3.3"
multidict = "==6.0.5"
requests-toolbelt = "1.0.0"
uvicorn = "==0.27.0"
web3 = "==6.1.0"
psutil = "^5.9.8"
pyinstaller = "^6.8.0"
aiohttp = "3.9.5"
deepdiff = "^8.0.1"

[tool.poetry.group.development.dependencies]
tomte = {version = "0.2.17", extras = ["cli"]}


[tool.poetry.group.dev.dependencies]
pytest = "*"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--cov=operate --ignore=scripts/test_e2e.py --ignore=scripts/test_staking_e2e.py --ignore=operate/data/contracts/uniswap_v2_erc20/tests/test_contract.py"
testpaths = [
    "./tests/",
]
norecursedirs = ["./operate/*", "scripts/*"]
